delete outdated drivers, change pushcolors lenght to 32 bit

+++ lib/libesp32_eink/M5EPD47/src/m5epd4in7.cpp
@@ -302,7 +302,7 @@ void M5Epd47::setAddrWindow(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1) 
 
 static inline void lvgl_color_swap2(uint16_t *data, uint16_t len) { for (uint32_t i = 0; i < len; i++) (data[i] = data[i] << 8 | data[i] >> 8); }
 
-void M5Epd47::pushColors(uint16_t *data, uint16_t len, boolean not_swapped) {
+void M5Epd47::pushColors(uint16_t *data, uint32_t len, boolean not_swapped) {
 
   nswapped = not_swapped;
